pyproject.toml
[tool.poetry]
name = "tennis_sql"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
sqlalchemy = "^1.3.18"
pytest = "^6.0.1"
pymysql = "^0.10.0"
cryptography = "^3.0"
pandas = "^1.1.0"
pydantic = "^1.6.1"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
arq = "^0.19"
requests = "^2.24.0"
python-dotenv = "^0.14.0"
pytest-dotenv = "^0.5.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"